fix:修复付费查看内容价格为整型会导致0元购的问题

+++ app/Plugins/Topic/src/Models/ShortcodePaidPost.php
@@ -35,5 +35,5 @@ class ShortcodePaidPost extends Model
     /**
      * The attributes that should be cast to native types.
      */
-    protected array $casts = ['id' => 'integer', 'post_id' => 'integer', 'amount' => 'integer', 'created_at' => 'datetime', 'updated_at' => 'datetime'];
+    protected array $casts = ['id' => 'integer', 'post_id' => 'integer', 'amount' => 'float', 'created_at' => 'datetime', 'updated_at' => 'datetime'];
 }
